requirements.txt
psycopg2-binary==2.7.5
requests==2.19.1
Scrapy==1.5.0
